[automation/python] Add Workspace.outputs
